// src/pages-fsd/ArticlesPage/model/types/articlesPageSchema.ts
import { EntityState } from '@reduxjs/toolkit';
import {
  Article,
  ArticleView,
  ArticleSortFields,
  ArticleType,
} from '@/entities/Article';
import { SortOrder } from '@/shared/types/sort';

// наследуемся от внутреннего типа redux, которое добавляет ids: string[] и entities с нашими типами комментов
export interface ArticlesPageSchema extends EntityState<Article, string> {
  // указывает, инициализировался у нас стейт или нет (чтоб заново его не инициализировать, когда стейт не удаляем из редакса)
  _inited: boolean;
  error?: string;
  // показывает, загрузили мы все статьи или есть еще
  hasMore: boolean;
  isLoading?: boolean;
  limit: number;
  order: SortOrder;
  // данные пагинации
  page: number;
  search: string;
  sort: ArticleSortFields;
  type: ArticleType;
  // фильтры
  view: ArticleView;
}